+++ sdk/identity/Azure.Identity/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.6.0-beta.1 (Unreleased)
 
 ### Features Added
-- `EnvironmentCredential` now supports certificate subject name / issuer based authentication with `AZURE_CLIENT_CERTIFICATE_ENABLE_SNI` environment variable.
+- `EnvironmentCredential` now supports certificate subject name / issuer based authentication with `AZURE_CLIENT_SEND_CERTIFICATE_CHAIN` environment variable.
 
 ### Breaking Changes